% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unobankproject.R
\name{bin_name}
\alias{bin_name}
\title{bin_name}
\usage{
bin_name(x)
}
\arguments{
\item{x}{Number}
}
\value{
Bins
}
\description{
Create bin's in increments
}
\details{
Calculate the percentage of concordant and discordant pairs for a given logit model.
}
\examples{
bin_name(100)
}
\author{
Arun Ranganathan \email{aranganathan@unomaha.edu}
}
