% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unobankproject.R
\name{pct}
\alias{pct}
\title{pct}
\usage{
pct(x)
}
\arguments{
\item{x}{Number}
}
\value{
table
}
\description{
Create pct's in a table
}
\details{
Calculate percentage/frequency.
}
\examples{
a<-rbind(sample(c(1:10),20, replace = T))
pct(a)
}
\author{
Arun Ranganathan \email{aranganathan@unomaha.edu}
}
