% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unobankproject.R
\name{pct_freq}
\alias{pct_freq}
\title{pct_freq}
\usage{
pct_freq(x)
}
\arguments{
\item{x}{is a vector of numbers or characters. This could take a column as well as a vector.}
}
\value{
table in a kable format using knitr package
}
\description{
Create percentage table and shows the result in a markdown format
}
\details{
This function takes a vector as input and creates their frequency and distribution. 
There are two columns output from this pct function. First column "count" shows number of occurrences of a string or a number
Second variable "Percentage" shows the percentage frequency of this number.
}
\examples{
a<-rbind(sample(c(1:10),20, replace = T))
pct(a)
}
\author{
Arun Ranganathan \email{aranganathan@unomaha.edu}
}
