% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unobankproject.R
\name{vif_glmrepeat}
\alias{vif_glmrepeat}
\title{vif_glmrepeat}
\usage{
vif_glmrepeat(dataset, depvariable, threshold)
}
\arguments{
\item{dataset, }{}

\item{dependent_variable}{}

\item{VIF_threshold}{}
}
\value{
variables that are below VIF threshold
}
\description{
Variable reduction method using VIF threshold
}
\details{
This function takes dataset, dependent variable, and threhold. 
Variance inflation factor is calculated using general linear model
Removes highest VIF variable and run glm again until we reach threshold 
Make sure to send clean data before running this function
}
\examples{
glmrepeat(german_credit, german_credit$Creditability, 2)
}
\author{
Arun Ranganathan \email{aranganathan@unomaha.edu}
}
